+++ tests/testthat/test-is_locatome_line.R
@@ -1,8 +1,6 @@
 test_that("use", {
-  expect_true(is_locatome_line("oooooMMMMMiiiii"))
-  expect_false(is_locatome_line(">5H2A_CRIGR"))
-  expect_false(is_locatome_line("5H2A_CRIGR"))
-  expect_false(is_locatome_line(NA))
-  expect_false(is_locatome_line(NULL))
-  expect_false(is_locatome_line(""))
+  expect_warning(
+    is_locatome_line("oooooMMMMMiiiii"),
+    "'is_locatome_line' is deprecated, use 'is_topology_sequence' instead"
+  )
 })