+++ tests/testthat/test-is_tmhmm_result.R
@@ -1,16 +1,6 @@
-test_that("use, result from run", {
-  if (!is_tmhmm_installed()) return()
-
-  fasta_filename <- system.file("extdata", "tmhmm.fasta", package = "tmhmm")
-  tmhmm_result <- run_tmhmm(fasta_filename)
-  expect_true(is_tmhmm_result(tmhmm_result))
-  expect_true(is_tmhmm_result(tmhmm_result[1:2]))
-  expect_false(is_tmhmm_result(tmhmm_result[1]))
-  expect_false(is_tmhmm_result(tmhmm_result[2]))
-})
-
-test_that("use, fake input", {
-  expect_false(is_tmhmm_result(NA))
-  expect_false(is_tmhmm_result(NULL))
-  expect_false(is_tmhmm_result(""))
+test_that("deprecated", {
+  expect_warning(
+    is_tmhmm_result(c(">A", "imo")),
+    "'is_tmhmm_result' is deprecated, use 'is_tmhmm_result' instead"
+  )
 })